requirements.txt
black==21.9b0
blinker==1.4
click==8.0.1
colorzero==2.0
Flask==2.0.1
Flask-Mail==0.9.1
gpiozero==1.6.2; platform_machine == 'armv7l'
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mypy-extensions==0.4.3
numpy==1.21.2
opencv-python==4.5.3.56
pathspec==0.9.0
picamera==1.13; platform_machine == 'armv7l'
platformdirs==2.4.0
pygame==2.0.1
python-dotenv==0.19.0
regex==2021.9.24
RPi.GPIO==0.7.0; platform_machine == 'armv7l'
tomli==1.2.1
twilio==7.1.0
typed-ast==1.4.3
typing-extensions==3.10.0.2
Werkzeug==2.0.1
zipp==3.6.0
